pyproject.toml
[project]
name = "gita-guru"
version = "0.1.0"
description = "A Streamlit application for Bhagavad Gita study and audio management"
authors = [
    {name = "Gita Guru Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "supabase>=2.8.0,<3.0.0",
    "gotrue>=2.8.0,<3.0.0",
    "httpx>=0.27.0,<0.28.0",
    "streamlit==1.47.0",
    "python-dotenv==1.0.1",
    "pandas==2.2.1",
    "numpy==1.26.4",
    "audio-recorder-streamlit==0.0.10",
]

# No build system needed for this Streamlit application

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
